changes made by siebert.
- You can add hosts, services, categories, protocols on the html interface -
no error checking, no error feedback. its on the todo list.
- Page refreshes every 60 secondes - bit of an ugly javascript but it does
the job, one might just move it to a better suitable place.
- Changes on the html interface updates the revision. I added a class method
to Cofiguration to handle this.

The css still needs some tweaking, but is good for now.

A patch for rails is implemented to prevent sqlite3 'too many open files' error.