Add a new field to the Note record to keep track of the associated node id.
When the user clicks on add node an Ajax call is made, the new note parameters are passed as HTTP POST (this should change to a RESTful aproach in the future).

When the selection changes in the tree, the selected node is stored in a new property of the NotesBrowser: +selectedNode+. This is used when creating a new note so we know to which node it is associated.