Add a new polling method (+checkrevision+) that grabs the value of the latest revision in the server and compares it with the local value in dradis.revision.
If there is a mismatch a new status message is presented urging the 
user to refresh.

There is a catch here, say we modify something, then we are the authors
of the change that lead to the latest increase in the revision number
but we already have the latest stuff in our client side. We want to 
prevent the status message from appearing if we made the last change.

In order to do so, we increase our local value for revision after a
successful Ajax call. This will keep both values in synch.