Change a bit the format in which the Notes are expected. The rdoc documentation already specified the right format:
#[<field #1 name>]# 
<field contents #1>

#[<field #2 name>]
<field contents #2>
 ...

For example:
 
#[Title]#
Insecure Cookie Configuration

#[Description]
Several flags that add an extra layer of security to HTTP cookies were 
found not to be in use by the application. 

The Secure flag was not set. 

The HTTPOnly flag was not set. 

The code (line #92) will split the note into a Hash like 
{ field1 => value1, field2 => value2 } And this will be used to populate the
template.