Replace the importsourcesDS data store for a new approach, we subclass the standard ComboBox to abstract all the JsonStore configuration.
In order to be able to access the different combos in the pannel we create
a new property: +fields+ and assign each item of the panel to it so we can
later do this.fields.sources, this.fields.filters, etc. This is useful for
instance to reload a specific data store.