Bring the latest changes from the trunk to the release branch:
The original idea of The Brain was to provide a centralised framework for ajax
requests.

We finally got our heads around how to implement this, and the dradis.ajax.request
method can be used to access Ext.Ajax. The idea is that this abstraction layer will
update the status bar with information and will also execute any callback passed
by the invoker once the Ajax call returns.

This is just a way of doing this and we will see whether it is interesting and 
useful or we should rather stick to the standard Ext.Ajax.


See original commit:
http://dradis.svn.sourceforge.net/viewvc/dradis?view=rev&revision=752