It is not uncommon that import/export plugins require the addition of new ActiveRecord objects to the database. For instance the WordExport plugin needs to create a new category "WordExport ready".
Plugin writers will add this code to their init.rb and when the server 
starts the code will be executed. The problem with this approach is that
plugins are also loaded when runing 'rake db:migrate'. If a plugin tries to
create an object before the table has been created then the process crashes.

In the environment.rb we can determine if we are in a rake process and prevent
plugins from loading.