Replace the <wx:sub-section id="vulntitle"> with a dummy tag with the same id.
This XML entity is not valid WordXML, however it is required due to the way in 
which the report is generated. The WordExport::Processor will add a paragraph 
containing the vulnerability title as a children of the XML element whose id is 
equal to "vulntitle". Fortunately enough Word is not bothered about this little 
hack.