Merging r1175 through r1185   from /branches/experimental/2009-06_refactorNodesTree-etd   into the trunk/
The new code of the NodesTree code is easier to read, and is better structured.
The pattern is similar to that used in the Importer and the AttachmentsViewer.

Additionally the new structure makes it easy the addition of event throwing 
code so the rest of the application can be notified of events originated in 
this widget.

Nevertheless the Ajax interactions of the NodesTree are awful and we should 
address it at a later stage.