Implement a new Export action in the Project Management plugin: metaserver_commit
It will pack the current project into a Zip archive and send it to the 
Meta-Server. Depending on the mode of operation (a new project vs. a project 
checked out of the MetaServer) a new project will be created in the Meta-Server
or a new Revision will be attached to the original Project.

Once the content of the project are zipped, the package is base64 encoded and
posted to the Meta-Server using ActiveResource's .post() method.