The truth is that people downloading the repository need these files in place in order to run the application.
The plugins take care of people installing the plugins through ./script/plugin
through the install.rb and uninstall.rb hooks. However, when someone checks out 
a revision from the repo, the file needs to exist so they can use it.

We are following the same convention we used for the database.yml config file,
provide a template (.template) and ignore the YAML file (.yml) once it exist.