Include the bare minimum plugins when running Rake.
This is to prevent plugins from interfering with the migrations as discussed 
in [i]. However, we need the project_management (and acts_as_tree) plugin to 
create the project package (.zip) used to create a backup.

By default dradis:reset is going to create a backup of the existing project.

[i]
http://dradis.svn.sourceforge.net/viewvc/dradis/server/trunk/config/environment.rb?view=log#rev874