Start working on the new Upload Manager. The idea is to improve the user experience when uploading results from other tools.
Right now the process is not very interactive and doesn't provide enough
feedback on what's going on. We are going to split the process into two stages:

* Upload the file
* Process the file

This will make it easier to detect and debug problems.


We will use a clean layout and some slick CSS from the flutie project [i] to 
get us started.

[i]
http://github.com/thoughtbot/flutie