Add a new Log model that will be used to log debug traces from background processing jobs into the database. This required because we want the background job to add logs to the DB and the web front end to poll for new logs.
This is ActiveRecorde backed although it is not backwards incompatible. The 
traces will not be included in project exports either.

The Log class has two main attributes: @uid and @text (we will probably add a 
@log_level later). Each file upload gets a unique id assigned so the backend job
and the frontend can identify the relevant log entries.

Currently the Log class only has two methods @write@/@debug@ that accepts an argument
or a block and uses it as a text for the entry and @read@ that pretty prints the log
entry.