This is a custom Delayed Job class. Read more about it at:
http://github.com/collectiveidea/delayed_job/

The only thing it does is create a logger with the unique id associated with
the current upload and call the corresponding Dradis upload plugin for
processing.

Due to the way delayed_job works, we cannot pass a complex object such as an
Attachment to the plugin, and since the only thing that plugins really need
to know is the attachment's path we will pass it as a string and refactor the
upload plugins so they expect a path instead of an Attachment instance.