The Upload#status action will be used to poll for new log entries for a given upload.
The parameters passed are the unique id of the upload and the id of the last log
entry already processed.

On the status.js.erb template we will render the logs/log partial for every new
log entry returned by the controller. We also keep the console's scroll bar down
to so users can see the last log entry appended.