Add a new action, Upload#parse. So far we have split the original upload process in two steps. First we upload (Upload#create) and the we parse the file.
The idea is that Upload#parse will enqueue the job to be process in the
background so the interface doesn't have to be waiting while it is being processed
without knowing what is going on.

See ./lib/upload_processing_job.rb for details but this action only takes the
name of the upload plugin, the attachment filename and unique id and enqueues
the job using delayed_job.

At the moment the view does nothing. After the upload is complete nothing needs 
to be done.