Add a new data store for the available Import plugins to the plugins manager.
At the end of the day, the list of plugins doesn't change after server boot so 
there is no need to refresh this store again and again. We will populated at the
beginning as the Upload and Export list and then pass a reference to it to the
dradis.importer.Panel object.

This means several dozen Ajax calls omited per session.