In Rails 3 ActiveRecord was broken in two layers, ActiveModel and ActiveRecord. As a result we need to adjust our fake ActiveRecord Attachment class to behave as the framework is expecting.
This model_name() method is used when you pass an instance of Attachment to the
render() method inside a view. ActionView tries to determine the path to the
corresponding partial automagically using the ActiveModel::Name object returned
by model_name()

We may need to consider re-implementing Attachment as full ActiveModel-compliant
or maybe think about Paperclip or something similar.