By default include RedCloth, now that we have the first-time content (r2085) we want people to have note formatting support.
The catch is that even when a binary version of RedCloth is available for Windows
users if we use a simple "gem 'RedCloth', :require => 'redcloth'" in the Gemfile
Bundler won't be able to find it. We will need to rely on the installer to get 
RedCloth into the system.

References:
http://groups.google.com/group/ruby-bundler/browse_thread/thread/892d56770fb7de5f