Remove the first time initializer. What this code was doing is create datbase objects the first time Dradis was launched after installation.
Although this was using Node::connected? to try to detect whether the migrations
had already been run, it was causing some issues with the latest SQLite3 ruby 
gem (1.3.2) in Windows (connected always returned @true@).

At the end of the day what the code is doing is adding new objects to the DB and
the only condition is that all the migrations have been run. This makes it very
suited to @db/seed.rb@.

I have moved the _first time_ related objects to @db/seed.rb@ and the code only
runs if the magic file (@config/first_login.txt@) doesn't exist.