Jason Garber from the RedCloth project was very helpful in pushing cross-compilation of the RedCloth gem.
Instead of the old MS VisualStudio compatible gem we used to use in the past 
they are now pushing a mingw32 binary version. The old binary was linked against
the Ruby 1.8 dll and wouldn't work with Ruby 1.9. The new MinGw32 is compatible
with the One-click installer.

I was planning to cross-compile the gem myself and include it in the installer
however, all this changes have already been pushed to the RedCloth repository, 
so we can use their version of the binary gem, we just need to use the --pre
switch.