Bj uses a cleaner interface to job descriptions than delayed_job used.
We no longer need to define classes with methods, the way in which our background
jobs are going to be run is using the rails _runner_. For instance:

ruby script/rails runner lib/upload_processing_job.rb NmapUpload "/tmp/nmap.xml" 1299859911740

So all the Rails environment will be available and we are just running the 
@lib/upload_processing_job.rb@ script as a stand-alone app. The parameters are 
used by the script to identify the import plugin, the file and a unique ID of 
the job.

The script will call the @import@ method of the plugin passing the file and a 
custom logger (see @app/models/log.rb@) that will store log traces in the DB 
associated with the job ID.

On the front-end, we will make Ajax calls to pull all this log entries for the
specific job.