We are going to change the way in which UploadController communicates with the upload plugin. Instead of passing an Attachment object representing the uploaded file, the controller will send a string with the path of the upload.
Upload plugins need to be aware of this fact se we perform a check on the
plugin's version number. If the version is lower than 2.7.0 there is a good 
chance the plugin is expecting an Attachment object.

This is mainly for the benefit of those creating custom upload plugins. We will
update all the plugins that come with the framework to expect a String instead
of an Attachment object.