Use the class object Attachment::AttachmentPwd instead of a hard-coded value for the attachments.
If for some reason we have to modify the location in which we store attachments
(as we had to do for the Heroku deployment) it is best to have the value in a 
single location rather than hard-coded all over the place.

It will also make things easier for when we make the attachments folder a 
configurable option.