To use NodesController with the ExtJS client-side libs, we are going to follow a similar approach to what we do in NotesController.
The :node_id is only used inside the nested resources (see rake routes), when
performing operations against nodes, we have to use :id.

We are going to send an Ajax call and the easiest way to send the data in the
request is through a JSON-encoded value in the :data parameter (this is also
the default approach in ExtJS REST stores).