Replace logger .debug() calls with .info() calls in the upload plugins
Full details can be found in issue #3 in the tracker [i] but the gist of it is
that Rails.logger wouldn't log traces under info() in production mode.

Our solution was to use @@logger.level = Logger.DEBUG to try to output the
traces but the truth is that most of the times it would make sense to use info()
instead of debug().

The other point is that when running through the Upload Manager, uploads are
processed in the background (using bj) and a custom logger (app/models/log.rb)
that logs the traces into the database (instead of STDOUT) is provided to the
plugins. To keep things simple, this logger doesn't support levels - i.e. level=()

[i]
https://github.com/dradis/dradisframework/pull/3
(cherry picked from commit 39d3e4bbb3acb1a408a8d4d0c8781bf9419951a6)