Replace :host, :port & :path settings w/ :rest_url
The old hard-coded "https://" approach was not too flexible if for some reason
Vuln::DB was deployed over HTTP instead of HTTPs.

With the new vulndb:rest_url we have full control of the ActiveResource's site
property.

A change in filters.rb also ensures that the latest setting is used each time.

This commit closses #7, #10