Provide more space and less Ajax to the Note Editor
When editting long notes the old editor window (horizontal split of editor and
preview) was a bit uncomfortable. We are splitting the the editor and preview
widgets in two tabs, so you can focus on editing or previewing your work.

The changes required:

* Switch from a Border layout Panel to a TabPanel (each widget provides a Title
attribute that will be used in the corresponding tab)

* Instead of updating the preview widget on every `keypress` just update when
the user clicks on the _Preview_ tab.

More info at:
http://sourceforge.net/mailarchive/forum.php?thread_name=4DD53FE2.1090401%40nomejortu.com&forum_name=dradis-devel