New updateNode Ajax method in dradis.NodesTree
Instead of making an Ext.Ajax call every time we need to update a node within
the class, create a helper method that lets us send updates to the server (that
will be processed by NodesController#update).

Make use of the new updateNode method after the user changes the label of a
node.