New Metasploit controller
We are going to transform the Metasploit Import plugin into a Rails engine with
its own views and actions. This is the first step, two actions have been
defined:

* _index_ to render the main layout that will present all the info available in
the remote Metasploit instance.

* _hosts_ will use the Msf RPC client to pull hosts and services from the
remote Metasploit. At the moment it is just a skeleton.

The MetasploitController inherits from AuthenticatedController, a new helper
class that buffers plugins from the authentication implementation.