New AuthenticatedController to require auth
This controller can be a handy one to inherit from if we are building
functionality that requires the user to be authenticated.

It isolates the code from the authentication implementation as we will be able
to change the details of the filter without the inheriting classes noticing.

Any third-party code (e.g. plugins) should inherit from this class instead of
calling the authentication filters directly.

It also sets the new default layout to whatever 'dradis3' will look like (at
the moment this is pretty much the clean layout of the Configuration and Upload
managers)