New configuration setting: :parent_node
This is the label of parent node for every node and note created by the plugin.
The new plugin will use the new node types to create 'host' nodes. Once a host
node is created if further information needs to be added to it, it will be
reused.

This differs from the previous implementation that did create a new node (with
the same label every time the information on the same IP was added).