Upload Manager: Process small attachments in line
Since the background processing of jobs takes some extra time (e.g. enqueue,
process, polling, etc.) it doesn't make sense to go through it if the uploaded
file is so small that it would only introduce a small delay to process it in
line.

For the time being we have a hard-coded threshold of 1MB, ideally this should
be a configurable setting.