Remove duplicated Thor password reset tasks
There was a dradis:reset:password and a dradis:setup:password it seems that
leaving dradis:reset:password makes more sense (name-wise). However I think it's
better to avoid passing the password as an argument (cached in user command
history, etc.) so the new task asks the user for password and confirmation and
then stores the SHA512 hexdigest of it.

This commit closes #43