Use RevisionObserver to create an audit log
In order to have smarter Ajax polling, we need a better audit log of the changes
made to the repo including identifying who made what change.

For instance, you would not be interested in updates of events you created
because the client-side GUI already took care of refreshing the relevant widgets.

We've added a virtual :updated_by attribute to the Note model (assigning it in
the NotesController) so we can keep track of the user that made the change.

Afterwards in RevisionObserver we create an audit log entry (:uid => 0) that
contains this :updated_by information. Later we can make the client-side polling
to request audit logs (not belonging to you) to update the GUI.