New LogsController to access the audit log
This controller is used by the Ajax poller to retrieve changes made by other
users.

Only Logs#index.json is defined. It returns all audit logs (:uid => 0) that are
associated with *other* users. This is because the current user's interface was
already updated in the browser before the actions reached the server and there
is no need for the interface to parse them again.