New method in NotesBrowser to enqueue a refresh
The .markDirty() method can be used by other widgets to notify the NotesBrowser,
that it has to refresh its contents as soon as possible. Two things can happen:

* If the user is editing / adding a note (editor window open), the
this.dirtyView property is set and as soon as the editor is closed (through the
'save', 'cancel' and 'hide' handlers) the data store is reloaded.

* If the user is not editing, then a refresh is triggered straight away.