New Importer: more efficient status polling
Instead of queueing a new status update just after making the Ajax call inside
`updateConsole()` in upload.js we wait until we receive a response from the
server (in app/views/upload/status.js.erb) to enqueue the next call.

This also has the benefit that we can delay a bit longer the request if the
previous call didn't contain any new logs.