Update first-time and what's new documentation for 2.8
This should have been changed before creating the git tag, but it slipped through
the cracks. I've added it to the _Preparing a new release_ wiki page [i] so it
doesn't happen again.

Instead of deleting the tag, or creating a new 2.8.1, the 2.8.0 packages contain
this changes and now we're adding them to the repo. Somone cloning v2.8.0 will
not have the changes, but these is non-critical information.

[i]
https://github.com/dradis/dradisframework/wiki/Preparing-a-new-release