Update the NessusUpload plugin to use Nokogiri
Instead of using the old REXML-based Nessus::Parser switch to a Nokogiri-based
approach.

In order to carry on using the custom ERB templates for hosts and report items
we need to make use of the Nessus::Host and Nessus::ReportItem helper classes
that hide the complexity of the XML format behind a friendly Ruby interface.