Do not create plugin categories during start up
The plugin's init.rb file may be loaded before the database is ready. For
instance when running RSpec for the first time, the spec Task depends on some
other tasks that end up loading the environment (and hence init.rb) before the
migrations are run.

If this happens, we will get a SQL error.

To avoid it, we create the Category once the plugin is invoked (i.e. when the
user uploads a file).