Convert the 'configurations' layout into the asset pipeline style
This involves a few changes, we are going to take this opportunity to merge the
'configurations' and 'dradis3' layouts, but not just yet.

The first step has been to remove the multiple JS and CSS includes and move them
to the dradis3.js and dradis3.css.scss manifest files. All the CSS code that was
included in the layout has also been moved to the dradis3.css.scss file.

The HTML @head@ element has been cleaned up and the document has been DOCTYPEd
to HTML5.