Bj: Make Bj compatible with Rails 3.2
In a nutshell, ActiveRecord 3.2 requires each class to be DB-backed. The
problem is that Bj::Table (supperclass of Bj::Table::Job, and subclass of
ActiveRecord) is not DB-backed.

When AR tries to configure Bj::Table::Job, it goes up the hierarchy to
Bj::Table and tries to configure it, but it fails because there is no DB table
or columns.

With this path we create a dummy 'bj_tables' DB table that solves the problem.

See the details in:
  ./vendor/plugins/bj/lib/bj/table.rb