New Evidence model
Evidence is what describes each specific instance of an Issue. It ties
the general problem description represented by the Issue (e.g. what SQL
injection) is, with the details specific to this case (e.g. for host X, and URL
Y, and parameter Z).

Each Evidence ties 1 Issue with 1 Node.