Remove local flag from bundle install
the bundle install command had the --local switch added, meaning that it would not connect to rubygems to try and install gems.  As some of the gems are not in the vendor/cache it was failing, causing problems with the install.  

So unless there's a specific reason to retain --local, this change suggests removing it.